feat(cli): Moved cli to keyweaver-cli
