--- .cobra.yaml
@@ -1,2 +1,0 @@
-author: 0siriz <0siriz@proton.me>
-license: MIT
