+++ cmd/keyweaver-cli/main.go
@@ -1,0 +1,9 @@
+package main
+
+import (
+	"github.com/0siriz/keyweaver/cmd/keyweaver-cli/cmd"
+)
+
+func main() {
+	cmd.Execute()
+}
